requirements.txt
asgiref==3.9.0
bcrypt==4.3.0
click==8.2.1
Django==5.2.4
djangorestframework==3.16.0
joblib==1.5.1
nltk==3.9.1
regex==2024.11.6
sqlparse==0.5.3
tqdm==4.67.1
